requirements.txt
pandas
numpy
scikit-learn
matplotlib
requests
python-dotenv